refactor(modal): extract controller modules
