+++ README.md
@@ -18,8 +18,9 @@ Inside of your Astro project, you'll see the following folders and files:
 /
 ├── public/
 ├── src/
+│   ├── controllers/
+│   │   └── modalController.ts
 │   ├── components/
-│   │   └── ModalController.ts
 │   └── pages/
 │       └── index.astro
 └── package.json
@@ -27,7 +28,7 @@ Inside of your Astro project, you'll see the following folders and files:
 
 Astro looks for `.astro` or `.md` files in the `src/pages/` directory. Each page is exposed as a route based on its file name.
 
-There's nothing special about `src/components/`, but that's where we like to put any Astro/React/Vue/Svelte/Preact components. Modal behavior is encapsulated in `src/components/ModalController.ts`, an Alpine.js controller used on the index page.
+There's nothing special about `src/components/`, but that's where we like to put any Astro/React/Vue/Svelte/Preact components. Modal behavior is handled by controller modules in `src/controllers/` and composed by `modalController.ts` for use on the index page.
 
 Any static assets, like images, can be placed in the `public/` directory.
