+++ src/test/modalController.test.ts
@@ -1,9 +1,10 @@
 import { describe, it, expect, vi } from 'vitest'
-import ModalController from '../components/ModalController'
+import createModalController from '../controllers/modalController'
+import { subscribeToModalEvents } from '../controllers/eventSubscriptions'
 
 describe('ModalController', () => {
   it('resets state on closeModal', () => {
-    const controller = ModalController()
+    const controller = createModalController()
     controller.isModalOpen = true
     controller.currentProject = {
       id: '1',
@@ -24,7 +25,7 @@ describe('ModalController', () => {
   })
 
   it('calculates next and previous slide indices', () => {
-    const controller = ModalController()
+    const controller = createModalController()
     controller.currentProject = {
       id: '1',
       title: 'Test',
@@ -40,5 +41,77 @@ describe('ModalController', () => {
 
     controller.prevSlide()
     expect(controller.changeSlide).toHaveBeenCalledWith(2)
+  })
+
+  it('openModal loads project data', async () => {
+    const controller = createModalController()
+    controller.preloadImage = vi.fn().mockResolvedValue(new Image())
+    ;(window as any).portfolioProjects = [
+      {
+        id: '1',
+        title: 'Test',
+        description: '',
+        audience: '',
+        slides: [{ image: { src: 'img.jpg' } }],
+      },
+    ]
+
+    await controller.openModal('1')
+
+    expect(controller.isModalOpen).toBe(true)
+    expect(controller.currentProject?.id).toBe('1')
+    expect(controller.isInitialLoad).toBe(false)
+    expect(controller.isContentVisible).toBe(true)
+  })
+
+  it('changeSlide updates slide index', async () => {
+    vi.useFakeTimers()
+    const controller = createModalController()
+    controller.currentProject = {
+      id: '1',
+      title: 'Test',
+      description: '',
+      audience: '',
+      slides: [{}, { image: { src: 'img2.jpg' } }],
+    }
+    controller.preloadImage = vi.fn().mockResolvedValue(new Image())
+
+    const promise = controller.changeSlide(1)
+    await vi.runAllTimersAsync()
+    await promise
+
+    expect(controller.currentSlideIndex).toBe(1)
+    expect(controller.isTransitioning).toBe(false)
+    vi.useRealTimers()
+  })
+
+  it('subscribes to modal events', () => {
+    const controller = createModalController()
+    controller.openModal = vi.fn()
+    controller.closeModal = vi.fn()
+    controller.nextSlide = vi.fn()
+    controller.prevSlide = vi.fn()
+    controller.closeZoom = vi.fn()
+
+    subscribeToModalEvents(controller)
+
+    document.dispatchEvent(
+      new CustomEvent('open-modal', { detail: { projectId: '1' } })
+    )
+    expect(controller.openModal).toHaveBeenCalledWith('1')
+
+    controller.isModalOpen = true
+    document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowRight' }))
+    expect(controller.nextSlide).toHaveBeenCalled()
+
+    document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowLeft' }))
+    expect(controller.prevSlide).toHaveBeenCalled()
+
+    document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+    expect(controller.closeModal).toHaveBeenCalled()
+
+    controller.isImageZoomed = true
+    document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+    expect(controller.closeZoom).toHaveBeenCalled()
   })
 })
